finished basic refactor of Account pages
