+++ src/components/pageComponentsAccount/Profile.tsx
@@ -1,0 +1,78 @@
+import React, {useState} from 'react';
+import axios from 'axios'
+
+import { GridHeader } from "@/components/commons/GridHeader";
+import { Box, Typography, Button, Alert, AlertTitle } from "@mui/material";
+
+import { FormReadOnly } from '@/components/forms/FormReadOnly';
+import { ProfileForm } from '@/components/forms/ProfileForm';
+import { ProfileFormData } from '@/lib/common/Types';
+import { AlertState, StatusOptions } from '@/lib/common/Types';
+import { UserPageProps } from "@/lib/common/Types";
+
+
+const initAlert = {
+    status: StatusOptions.ERROR,
+    message: "",
+    isDisplay: false,
+}
+const alertTimer = 1000;
+
+
+export const Profile = ({user, ...props}:UserPageProps) => {
+    const [alert, setAlert] = useState<AlertState>(initAlert);
+    const [isEdit, setIsEdit] = useState(false);
+    const [userSnapshot, setUserSnapshot] = useState(user);
+
+    const {username, first_name, last_name, telephone} = userSnapshot;
+    const data = [
+        {title: "Username", value: username},
+        {title: "First Name", value: first_name},
+        {title: "Last Name", value: last_name},
+        {title: "Telephone", value: telephone},
+    ];
+    const formReadOnlyElems = <FormReadOnly data={data} />
+
+    const handleSubmit = async (formData:ProfileFormData) => {
+        try {
+            const response = await axios.post('/api/forms/profile', {
+                id: user.id, ...formData
+            });
+
+            setUserSnapshot({...userSnapshot, ...formData});
+            setAlert({status: StatusOptions.SUCCESS, message: response.data.message, isDisplay: true});
+        } catch (error) {
+            console.log("🚀 ~ file: profile.tsx:40 ~ handleSubmit ~ error:", error)
+            setAlert({status: StatusOptions.ERROR, message: "Error updating profile.", isDisplay: true});
+        }
+
+        setIsEdit(false);
+    }
+
+    const formElem = <ProfileForm user={userSnapshot} handleSubmit={handleSubmit} />
+
+    const alertElem = alert.isDisplay? <Alert severity={`${alert.status}`}
+        onClose={ () => setAlert(initAlert) }
+    >
+        <AlertTitle>{alert.message}</AlertTitle>
+    </Alert> : <></>;
+
+    return (<Box>
+
+        <GridHeader >
+            <Typography component="span" variant="h5" ml={2} >
+                Profile
+            </Typography>
+            <Button variant="outlined" sx={{ marginRight: '0.5rem' }}
+                onClick={() => setIsEdit(!isEdit)}
+            >
+                {!isEdit? "Edit" : "Cancel"}
+            </Button>
+        </GridHeader>
+
+        {alertElem}
+
+        {isEdit? formElem : formReadOnlyElems}
+
+    </Box>)
+}
