+++ src/pages/account/index.tsx
@@ -1,97 +1,15 @@
 import React, {useState} from 'react';
-import axios from 'axios'
 
 import { GetServerSidePropsContext } from "next";
 import { isLoginRequiredServer } from "@/lib/auth";
 import { fetchUserByEmail } from "@/lib/mongoDB/userQueries";
-import { UserPageProps } from "@/lib/common/Types";
 
 import AccountWrapper from "@/components/layouts/AccountWrapper";
 import CenteredWrapper from "@/components/layouts/CenteredWrapper";
-import { GridHeader } from "@/components/commons/GridHeader";
-import { Box, Typography, Button, Alert, AlertTitle } from "@mui/material";
 
-import { FormReadOnly } from '@/components/forms/FormReadOnly';
-import { ProfileForm } from '@/components/forms/ProfileForm';
+import { UserPageProps } from "@/lib/common/Types";
 
-import { ProfileFormData } from '@/lib/common/Types';
-import { AlertState, StatusOptions } from '@/lib/common/Types';
-
-
-const initAlert = {
-    status: StatusOptions.ERROR,
-    message: "",
-    isDisplay: false,
-}
-const alertTimer = 1000;
-
-export default function Profile({user, ...props}:UserPageProps){
-
-    const [alert, setAlert] = useState<AlertState>(initAlert);
-    const [isEdit, setIsEdit] = useState(false);
-    const [userSnapshot, setUserSnapshot] = useState(user);
-
-    const {username, first_name, last_name, telephone} = userSnapshot;
-    const data = [
-        {title: "Username", value: username},
-        {title: "First Name", value: first_name},
-        {title: "Last Name", value: last_name},
-        {title: "Telephone", value: telephone},
-    ];
-    const formReadOnlyElems = <FormReadOnly data={data} />
-
-    const handleSubmit = async (formData:ProfileFormData) => {
-        try {
-            const response = await axios.post('/api/forms/profile', {
-                id: user.id, ...formData
-            });
-
-            setUserSnapshot({...userSnapshot, ...formData});
-            setAlert({status: StatusOptions.SUCCESS, message: response.data.message, isDisplay: true});
-        } catch (error) {
-            console.log("🚀 ~ file: profile.tsx:40 ~ handleSubmit ~ error:", error)
-            setAlert({status: StatusOptions.ERROR, message: "Error updating profile.", isDisplay: true});
-        }
-
-        setIsEdit(false);
-    }
-
-    const formElem = <ProfileForm user={userSnapshot} handleSubmit={handleSubmit} />
-
-    const alertElem = alert.isDisplay? <Alert severity={`${alert.status}`}
-        onClose={ () => setAlert(initAlert) }
-    >
-        <AlertTitle>{alert.message}</AlertTitle>
-    </Alert> : <></>;
-
-
-    return <>
-        <AccountWrapper>
-            <CenteredWrapper>
-                <Box>
-
-                    <GridHeader >
-                        <Typography component="span" variant="h5" ml={2} >
-                            Profile
-                        </Typography>
-                        <Button variant="outlined" sx={{ marginRight: '0.5rem' }}
-                            onClick={() => setIsEdit(!isEdit)}
-                        >
-                            {!isEdit? "Edit" : "Cancel"}
-                        </Button>
-                    </GridHeader>
-
-                    {alertElem}
-
-                    {isEdit? formElem : formReadOnlyElems}
-
-                </Box>
-            </CenteredWrapper>
-        </AccountWrapper>
-    </>
-}
-
-
+import { Profile } from '@/components/pageComponentsAccount/Profile';
 
 
 export async function getServerSideProps(context:GetServerSidePropsContext) {
@@ -122,4 +40,22 @@ export async function getServerSideProps(context:GetServerSidePropsContext) {
         },
     };
 }
+
+export default function ProfilePage({user, ...props}:UserPageProps){
+
+    return <>
+        <AccountWrapper>
+            <CenteredWrapper>
+                <Profile
+                    {...{
+                        user
+                    }}
+                />
+            </CenteredWrapper>
+        </AccountWrapper>
+    </>
+}
+
+
+
